Makefile: Don't pass -polly-opt-fusion=max if CLANG > 13.0.0
The command line option -polly-opt-fusion has been completely removed [1] in the latest LLVM toolchain (and in the latest Clang 14 build)

The cmdline option was removed previously [2], though the main switch remained because of a human error
which kept this option functioning in the polly optimizer. But in reality, it did nothing since then.
Now, it has been completely removed.

Let's use a condition to make compiler use this flag if clang's version is less
than 14.

The name of the option is misleading and has been renamed by isl to
"serialize-sccs". Instead of also renaming the option, it has been removed by llvm.

The option is still accessible using the following cmdline option:

[ -polly-isl-arg=--no-schedule-serialize-sccs ]

[1] llvm/llvm-project@cb879d0

[2] llvm/llvm-project@07e7cb9

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>